refactor: 컨트롤러 분리
- 도메인 단위로 컨트롤러를 분리합니다.
- 컨트롤러의 중복된 매핑을 제거합니다.
