+++ src/main/java/roomescape/presentation/page/AdminPageController.java
@@ -1,27 +1,29 @@
-package roomescape.presentation;
+package roomescape.presentation.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/admin")
 public final class AdminPageController {
 
-    @GetMapping("admin")
+    @GetMapping
     public String admin() {
         return "admin/index";
     }
 
-    @GetMapping("admin/reservation")
+    @GetMapping("/reservation")
     public String reservation() {
         return "admin/reservation-new";
     }
 
-    @GetMapping("admin/time")
+    @GetMapping("/time")
     public String time() {
         return "admin/time";
     }
 
-    @GetMapping("admin/theme")
+    @GetMapping("/theme")
     public String theme() {
         return "admin/theme";
     }
